+++ front/app/components/admin/GraphCards/PostsByTimeCard/usePostsByTime/typings.ts
@@ -7,14 +7,6 @@ import {
 
 export type QueryParameters = ProjectId & Dates & Resolution;
 
-export interface TimeSeriesResponseRow extends InputsCountRow {
-  first_dimension_date_created_date: string;
-}
-
-export interface InputsCountRow {
-  count: number;
-}
-
 // Hook return value
 export interface TimeSeriesRow extends TimeSeriesTotalRow {
   inputs: number;
